+++ changelog.md
@@ -30,6 +30,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Unicode escape sequences that are not 2 or 4 bytes long can now be specified with `\u{…}`
 - Change [`pop ◌`](https://uiua.org/docs/pop)'s glyph to make it look good in planet notation. Code using `;` will continue to work and will be formatted as `◌`.
 - [`un °`](https://uiua.org/docs/un)[`reduce /`](https://uiua.org/docs/reduce)[`multiply ×`](https://uiua.org/docs/multiply) now gives the prime factorization of a number
+- [`classify ⊛`](https://uiua.org/docs/classify) now works with [`under ⍜`](https://uiua.org/docs/under)
 - Completely remove the deprecated `unbox ⊔`
 ### Interpreter
 - The internal byte array type is now used in more places, which should improve performance a bit